Start migrating to the new questionnaire model
MVP: simply ask for free holidays in its own questionnaire, remove option for reporting normal holidays
